requirements.txt
geopandas==0.13.2
numpy==1.24.3
pandas==2.0.3
PyYAML==6.0.1
scipy==1.10.1
statsmodels==0.14.0
joblib==1.3.1
scikit-learn==1.3.0
matplotlib==3.7.2
seaborn==0.12.2
jupyter==1.0.0
notebook==7.0.2
pytest==7.4.0
pytest-cov==4.1.0
black==23.7.0
flake8==6.1.0
mypy==1.5.1
python-dotenv==1.0.0
tqdm==4.65.0
loguru==0.7.0
shapely==2.0.1
pyproj==3.6.0
fiona==1.9.4
rtree==1.0.1
networkx==3.1
folium==0.14.0
branca==0.6.0
